requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Bootstrap-Flask==2.5.0
Flask-Limiter==4.1.1
gunicorn==21.2.0
pillow==10.4.0
openpyxl==3.1.2
pandas==2.3.3
requests==2.31.0
SQLAlchemy==2.0.25
WTForms==3.0.1
psycopg2-binary
vercel-blob